[Justice Counts] Use DB Jurisdictions to Return Census Populations (Recidiviz/recidiviz-data#35556)
## Description of the change

This PR leverages saved census data in the `jurisdiction_population`
table to identify and surface jurisdictions associated with an agency.
You can test the changes
[[here](https://nichelletest---publisher-web-b47yvyxs3q-uc.a.run.app/agency/147/settings)]
for `Law Enforcement` on my playtesting branch.
[Here](https://nichelletest---publisher-web-b47yvyxs3q-uc.a.run.app/api/v2/agencies/147/published_data)
is the link to the API response.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e48924ad448c4f55e60da7adce7899382a6f5a6e