[Observations] Remove JSON attributes column from single-observation views (Recidiviz/recidiviz-data#35591)
## Description of the change

I've observed in my aggregated metric perf testing that BQ isn't
actually storing the result of `JSON_EXTRACT` calls that happen inside a
subquery and reusing those values.

For example, in a query like this:
```
SELECT
    COUNTIF(my_attribute = "A") AS a_count,
    COUNTIF(my_attribute = "B") AS b_count,
    COUNTIF(my_attribute = "c") AS c_count,
FROM (
     SELECT JSON_EXTRACT(event_attributes, "$.my_attribute) AS attribute
     FROM observations_table
)
```
... even though it looks like we'd be extracting `my_attribute` once and
then using it 3 times, under the hood it looks more like it's doing:
```
SELECT
    COUNTIF(JSON_EXTRACT(event_attributes, "$.my_attribute) = "A") AS a_count,
    COUNTIF(JSON_EXTRACT(event_attributes, "$.my_attribute) = "B") AS b_count,
    COUNTIF(JSON_EXTRACT(event_attributes, "$.my_attribute) = "c") AS c_count,
FROM (
     SELECT  event_attributes
     FROM observations_table
)
```

As a result, I'm pushing forward some work in this PR that I didn't
think we'd have to do right now, which is to stop packaging attributes
into JSON in the individual observation queries, and only package them
into JSON for the unioned `all_person_events` etc views that need to
have a shared schema for each observation.

This means that in my optimized aggregated metrics queries we'll not be
doing ANY JSON_EXTRACT calls, hopefully significantly reducing runtime.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34498

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 58ea18aaa2327ca2b128103691a768bdaf7a5513