[US_AZ] Add classification score to custody level raw text (Recidiviz/recidiviz-data#35543)
## Description of the change

This adds the raw classification score to the raw text of the
`custody_level` field on the `incarceration_period` entity. We still map
`custody_level` values based on the text description of the final
custody level as it appears on the classification report overall, rather
than based on the score itself, but the score will now be accessible in
the state database.

Test pipeline:
https://console.cloud.google.com/dataflow/jobs/us-west1/2024-11-26_09_58_18-18027942135711034832;graphView=1?project=recidiviz-staging

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35537

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ea9376a3e7eb380cb6459c17fc82bba869a0fbfd