[db] add mixin for using cloud sql proxy from manager classes (Recidiviz/recidiviz-data#35596)
## Description of the change

this PR adds a new class, `CloudSqlProxyMixin`, that handles connecting
either directly to a cloud sql db or via a proxy. i had considered doing
this w/in the `SessionFactory` class (see that approach in draft Recidiviz/recidiviz-data#35597,
will close in favor of this PR) but in the end felt like it made sense
to not overload `using_database` too much -- lmk if y'all prefer that
approach!

in the process of writing Recidiviz/recidiviz-data#35566, i wanted to be able to use the
existing metadata manager to acquire/release raw data resource locks;
however, since it is running locally, we can't really do that w/ how
things are currently set up. this PR is the way i thought to solve this
issue

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Prework for Recidiviz/recidiviz-data#35566

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 494c8db883bedbed95256e3a7035c46c0958f8c2