[Workflows] Add defaults for missing config fields (Recidiviz/recidiviz-data#35611)
## Description of the change

There were a handful of fields added to the config zod schema that were
not added to the webform. This caused the zod parser to reject them.

This PR addresses that issue by setting sane defaults for many of the
fields in the zod definition. This way, if the fields aren't explicitly
set, we can have usable values. I also moved some defaults set in the
form view to the zod definition. This way, we can reuse the parser in
multiple places when needed without needed to reimplement default values
everywhere.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4120a5a559b213b0f6f5dbd5fe5ebfc73f52da05