[US_ND] Hydrate Sentences V2 entities (Recidiviz/recidiviz-data#35479)
## Description of the change

This PR hydrates `StateSentence`, `StateSentenceLength`,
`StateChargeV2`, `StateSentenceGroup` with data from the Elite and
Docstars systems. It also hydrates `StateSentenceGroupLength` with data
from the Elite system.

Since these two systems are created by different organizations and are
not designed to link to one another, the external IDs for each of these
entities across systems are not consistent. This hydration relies
exclusively on the normalization logic that creates sentence inferred
groups to join information across systems.

In Elite, the data is structured in a way that lends itself much more
easily to being organized into sentences and sentence groups. In
Docstars, there is no clear concept of a sentence group; we construct
these ourselves based on logical inferences about how sentences are
imposed in ND. Even so, the structure is not necessarily intuitive.

### Docstars (supervision)
- The `state_sentence` entity will hold all components of a given
sentence. If there is a pre-trial period followed by a period of
probation, there will be one `sentence` object for each of those
components. This is how the information is stored in Docstars. Each
component will have its own `imposed_date` and end date. In this
scenario, the components would line up end-to-end.
- The `state_sentence_group` entity is a container for all of the
components that stem from a given court case. This means that in the
example above, the pre-trial component and the probation component will
be a part of the same `sentence_group`. This linking is done by the ID
of the court case that appears in the raw data with each individual
sentence component.
- Since there is no concept of a `sentence_group` in Docstars itself,
there are no dates associated with the overall sentence group stored in
Docstars; dates in Docstars only correspond to the individual components
that make up a sentence group. `StateSentenceGroupLength` is not
hydrated from Docstars raw data.

### Elite (incarceration)
- A sentence group is known as a "booking". A "booking" refers to one
incarceration stint. All charges associated with that incarceration
stint are attached to the ID for that booking (`OFFENDER_BOOK_ID`), and
all sentences associated with that booking (concurrent or consecutive)
are differentiable.
- The structure in this PR mirrors the existing structure in
`state_incarceration_sentence`, with the addition of
`state_sentence_group` that compiles information at the level of the
booking itself.
- This PR adds the `is_violent`, `is_sex_offense`, and `is_drug` flags
that were not previously hydrated to the `StateChargeV2` entity.

### Charges
Being able to link information from Elite to Docstars using sentence
inferred groups helps us to see the full picture of a person's
interaction with the justice system, but charges are duplicated across
these systems and cannot easily be deduplicated. Charges are stored
using different code systems in each system (NCIC codes in Elite, Common
Statute Codes or NCIC codes in Docstars), and each system has a distinct
reference file for translating those codes into charge descriptions.
They do not yield the same descriptions across systems; this is a known
issue in ND and is not slated to be fixed in the short term. A result is
that we cannot tell which charges correspond to a singular real-life
event across systems.

A practical impact of this is that we are unable to accurately tally how
many offenses a person has been convicted of in relation to a specific
sentence or time period.

### Tests
Testing added in this PR confirms the following:
- Sentence components stored in each system are correctly joined into a
single sentence inferred group when they share an `offense_date` and a
Recidiviz-generated `person_id`.
- There are no `SentenceGroupLength` objects that relate to sentences
that exist exclusively in Docstars. If there is a sentence that has
components in both Docstars and Elite (i.e. a person was sentenced to a
term of imprisonment followed by a period of probation), then there will
be a `sentence_group_length` object created only based on the
incarceration component of that sentence.

Successful sandbox ingest pipeline
[here](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-11-25_13_18_57-2066228608556865455;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fd3e5bab549005f746f63b8e7b29fa16256add0a